Split type definition into seperate files
