+++ types/firewall.ts
@@ -1,0 +1,7 @@
+export type FirewallFields = 'country' | 'continent' | 'asn' | 'ip' | 'hostname' | 'user-agent';
+export type FirewallOperators = 'equal' | 'not equal' | 'greater' | 'less' | 'in' | 'not in' | 'contain' | 'not contain';
+export interface FirewallOptions {
+  field: FirewallFields;
+  operator: FirewallOperators;
+  value: string | string[] | number | number[];
+}
