+++ types/optimization.ts
@@ -1,0 +1,8 @@
+export interface OptimizationOptions {
+  mirage?: boolean;
+  minify?: {
+    javascript?: boolean;
+    css?: boolean;
+    html?: boolean;
+  };
+}
